# Add in-place conversion to basic data format