@ccoffrin 2024-06-15
Shall we add a test for `make_basic_network!` and `make_basic_network` that confirms that one is destructive and the other is not?