@ccoffrin 2024-06-15
After that I'll be good to merge.  Feel free to add your name in the list of the readme if you like.